Add support to projec for Travis CI integration
With this change we are integrating the project with a continuous integration service (in this case [Travis CI](https://travis-ci.org/EnvironmentAgency/flood-risk-acceptance-tests).

This will ensure that changes we make to the project don't break the project. At this time we are excluding failing Cucumber scenarios from what we define as something that breaks the project, and instead focusing on the project fundamentals, that Quke is operating normally and Cucumber can be called, find our features and complete them without issue.

Also as it is CI we are also including running rubocop to ensure that contributions do not break any style and convention rules.

To do this we have had to

- add a `.travis.yml` config file
- add a `Rakefile` with some custom tasks
- add a dependency to the `Gemfile`
- add a feature and steps file for a scenario which does not require web access to run