Integrate project with CodeClimate
This change integrates the project with [CodeClimate](https://codeclimate.com), an online service that consolidates the results from a suite of static analysis tools into a single, real-time report.

Specifically we are adding a `.codeclimate.yml` file which is used to configure which tools CodeClimate actually analyses this project with.