Add first validation scenario and feature
The initial features and scenarios in this projct were ported from an existing project. During this process the scenario to check that a user is returned to the add exemption page when they remove their previously selected exemption was found in the local authority feature. However as you never even get to the org type it was thought to be clearer to move the scenario to its own feature named *Validation*.