Add initial dredging feature and scenarios
The initial features and scenarios in this project were ported from an existing one.

During this process the scenarios to test that a field which asks for approximate dredging length is only shown when a certain exemption is picked was found in the local authority feature. However as you never even get to the org type it was thought to be clearer to move the scenario to its own feature named *dredging*.