Add details on tags and their usage to README
To ensure consistent use of tags across the project, and also to provide a definition of when to use them for new users, this change adds details about them to `README.md`.

It also makes house keeping changes to the feature and scenario files to ensure they are consistent with the definitions we've set out.