Add check for @focus and @wip tags to .travis.yml
Having detailed in e627ae0 that we don't want temporary tags like @focus and @wip to be merged into **master** this change adds a custom step in the Travis CI build (via the .travis.yml config file) to check for them.

If it finds the tags it will fail the build.