Create distinct check your answers feature
Currently the scenario *Registration by a local authority* in [local_authority_registration.feature](features/local_authority_registration.feature) has additional content in its `I am a local authority` step.

This is a series of `expect()` statements that are checking the content of the **Check your answers** page. This change brings the scenario inline with the rest of the registration scenarios by moving this to a new **functional** feature that provides a focus for the **check your answers** page.