Add partners scenario to check answers feature
This change adds a scenario to the check answers feature which covers registering as a partnership.

Registering as a partnership if one of the user types known to cause a change in how the details are displayed in the **check your answers** page. This scenario covers this change, which is to display each of the partners added individually.