Update Quke gem to latest (plus others)
Updating the [Quke](https://github.com/EnvironmentAgency/quke) gem to take advantage of an update to its ability to work with proxy servers, specifically the option of specifying addresses in the config where the driver should NOT use the proxy to make a connection.

This will resolve an issue with getting this project running in our Jenkins build environment.