Update expect's in login steps to use SitePrism
Realised that we weren't making use of [SitePrism](https://github.com/natritmeyer/site_prism#testing-for-the-existence-of-the-element) when it came to testing for the existence of an element. Previously we would get the element but then check a property of it.

This change updates the syntax used in our back office login expect statements to use features built into **SitePrism** to do this kind of checking.

Doing this makes the steps more maintainable, and the logic simpler to read and understand.