Refactor BackOfficeHomePage to separate UI logic
This change makes a distinction between the functionality needed to set the root url for a test, which is either the front or back office home page, and interactions with it.

In the front office the root or home page is simply a check list with a continue button. In the back office though it can be different pages depending on the context; when not signed in its the login page, but when you are signed in it becomes the search page.

To make a clear distinction between the logic needed to set the root url at the start of a scenario, and that needed to interact with the various pages this change splits out the login UI logic into a new `logic_page.rb` page object (`search_page.rb` already follows this pattern). The `set_url` logic is left in `back_office_home_page.rb`.