Add scenario 2 reproduce duplicate regs in search
https://eaflood.atlassian.net/browse/RIP-359

This change adds a scenario which reproduces a situation where an exemption will appear twice in the back office search results because it has multiple organisation address listed as the primary.

There should only ever be one address linked to the record as the primary address, however by using the back links and choosing to re-enter the address manually we can demonstrate how 2 address records are created, linked to the organisation record (which in turn is linked to the enrollment record) resulting in the exemption appearing twice in search results.