Lock chromedriver-helper to 1.0.0 (#86)
First step in an attempt to fix broken automated acceptance tests in our non-production environments.

Having recently tried to run an updated version of this project in our pre-production environment we have found it to no longer work. The current belief is because the chromedriver version of the project is now no longer in sync with that installed on our Jenkins slave it can no longer connect and run.

This change locks the version of `chromedriver-helper` to one we know that works in our environments.