Lock Capybara to version 2.14.0 (#87)
In recent testing we have found that the scenario 'Add 3 partners then remove one' in `partnership.feature is failing because it is unable to confirm the javascript prompt that is displayed. When run on the line `page.accept_confirm` when called we get a `undefined method `[]' for nil:NilClass (NoMethodError)`.

The release notes for 2.14.1 list _Workaround broken system modals when using selenium with headless Chrome_ so it may be that this fix is actually a breaking change for us.

Therefore reverting the version back until we understand how to fix it.