Add ability to specifiy custom screen size
As of writing it was 10 months since the tests ran successfully in our environments. When attempting to confirm the service after some emergency patching of the environments we

- could not get the acceptance tests to connect to chrome (20a8ea7)
- errored when trying to confirm the JavaScript prompt after deleting a partner (35920c1)
- could not select the correct address (ac95bb8)

With those issues out of the way, and the versions fixed we then found we were experiencing the [same problem as the WEX acceptance test project](https://github.com/DEFRA/waste-exemptions-acceptance-tests/pull/59). Basically in our Jenkins CI environment if an element is not visible in the display port Capybara cannot select it.

So we have implmented the same fix, which is to add the ability to specify a custom screen size within the config.