Move config to its own folder for consistency
In the [waste-exemptions-acceptance-tests](https://github.com/DEFRA/waste-exemptions-acceptance-tests) the config files are stored in their own `config` folder and have a consistent format for the file names.

Though not critical, it splits the configuration into environment specific files. This means you are not having to keep commenting and uncommenting sections of a sourc controlled file. It also just makes things a little easier to navigate in the terminal, in your editor, and in the GitHub UI.

Finally, its moves this project another step closer to being consistent with the WEX acceptance tests, which will help when having to jump in and out of them.